tests: Split NetCDF engine tests into separate modules
