+++ tests/engines/netcdf4_engine/test_netcdf_dimension_converter.py
@@ -3,7 +3,7 @@
 import numpy as np
 import pytest
 
-from tiledb.cf.engines.netcdf4_engine import NetCDFDimensionConverter, open_netcdf_group
+from tiledb.cf.engines.netcdf4_engine import NetCDFDimensionConverter
 
 
 def test_dim_converter_simple(tmpdir_factory):
@@ -35,41 +35,3 @@ def test_dim_converter_unlim(tmpdir_factory):
         assert converter.name == dim.name
         assert converter.domain == (0, max_size - 1)
         assert converter.dtype == np.uint64
-
-
-def test_open_netcdf_group_with_group(tmpdir):
-    netCDF4 = pytest.importorskip("netCDF4")
-    filepath = str(tmpdir.mkdir("open_group").join("simple_dataset.nc"))
-    with netCDF4.Dataset(filepath, mode="w") as dataset:
-        with open_netcdf_group(dataset) as group:
-            assert isinstance(group, netCDF4.Dataset)
-            assert group == dataset
-
-
-def test_open_netcdf_group_with_file(tmpdir):
-    netCDF4 = pytest.importorskip("netCDF4")
-    filepath = str(tmpdir.mkdir("open_group").join("simple_dataset.nc"))
-    with netCDF4.Dataset(filepath, mode="w") as dataset:
-        group1 = dataset.createGroup("group1")
-        group1.createGroup("group2")
-    with open_netcdf_group(input_file=filepath, group_path="/group1/group2") as group:
-        assert isinstance(group, netCDF4.Group)
-        assert group.path == "/group1/group2"
-
-
-def test_open_netcdf_group_bad_type_error():
-    with pytest.raises(TypeError):
-        with open_netcdf_group("input_file"):
-            pass
-
-
-def test_open_netcdf_group_no_file_error():
-    with pytest.raises(ValueError):
-        with open_netcdf_group():
-            pass
-
-
-def test_open_netcdf_group_no_group_error():
-    with pytest.raises(ValueError):
-        with open_netcdf_group(input_file="test.nc"):
-            pass
